+++ src/QueryBuilder/DSL/Query.php
@@ -111,11 +111,14 @@ class Query implements DSL
      * common terms query.
      *
      * @see https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-common-terms-query.html
+     * @deprecated since version 7.1.3, use the "match()" method instead.
      *
      * @param float $cutoffFrequency percentage in decimal form (.001 == 0.1%)
      */
     public function common_terms(string $field, string $query, float $cutoffFrequency): Common
     {
+        \trigger_deprecation('ruflin/elastica', '7.1.3', 'The "%s()" method is deprecated, use "match()" instead. It will be removed in 8.0.', __METHOD__);
+
         return new Common($field, $query, $cutoffFrequency);
     }